recommendPoint 최적화

+++ src/main/java/com/koreaIT/demo/controller/UsrRecommendPointController.java
@@ -14,34 +14,29 @@ public class UsrRecommendPointController {
 
 	private RecommendService recommendService;
 	private Rq rq;
-
+	
 	UsrRecommendPointController(RecommendService recommendService, Rq rq) {
 		this.recommendService = recommendService;
 		this.rq = rq;
 	}
-
+	
 	@RequestMapping("/usr/recommendPoint/getRecommendPoint")
 	@ResponseBody
 	public ResultData<RecommendPoint> getRecommendPoint(String relTypeCode, int relId) {
 		return recommendService.getRecommendPoint(rq.getLoginedMemberId(), relTypeCode, relId);
 	}
-
-	@RequestMapping("/usr/recommendPoint/insertRecommendPoint")
+	
+	@RequestMapping("/usr/recommendPoint/doRecommendPoint")
 	@ResponseBody
-	public String insertRecommendPoint(String relTypeCode, int relId) {
-
+	public String doRecommendPoint(String relTypeCode, int relId, boolean recommendBtn) {
+		
+		if (recommendBtn) {
+			recommendService.deleteRecommendPoint(rq.getLoginedMemberId(),relTypeCode, relId);
+			return "좋아요 취소";
+		}
+		
 		recommendService.insertRecommendPoint(rq.getLoginedMemberId(),relTypeCode, relId);
-
 		return "좋아요 성공";
 	}
-
-	@RequestMapping("/usr/recommendPoint/deleteRecommendPoint")
-	@ResponseBody
-	public String deleteRecommendPoint(String relTypeCode, int relId) {
-
-		recommendService.deleteRecommendPoint(rq.getLoginedMemberId(),relTypeCode, relId);
-
-		return "좋아요 취소";
-	}
-
+	
 }
